Add user input validations for command arguments
I spent some time trying to figure out a clean looking way to do this,
and settled upon the commands knowing exactly what order the arguments
should be in and whether they are a number or letter (or no arguments).
This enabled me to validate the arguments in one place before continuing
on to the command case statement.