Output error if trying to colour outside of bitmap
The Bitmap model will raise an error if trying to colour a coordinate
that doesn't exist on the grid, then the error is rescued in
BitmapEditor and error is output.

In the Bitmap model this meant I had to stop using the `set_colour`
method when initializing the grid, as it would raise the
OutOfBoundsError immediately because the grid initially has no
coordinates.